moved SelfKnowledge inside barabasz.py
